requirements.txt
streamlit  
openai  
azure-storage-blob  
azure-cognitiveservices-vision-computervision  
msrest  
python-dotenv  
pillow  